Code cleanup & linting
